moved mock objects into their own directory and renamed mock event service to make obvious what it is
