+++ src/app/event-view/event-list/event-list.component.ts
@@ -1,7 +1,6 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 
 import { Event } from '../../shared/event.model';
-import { EventService } from '../../shared/event.service';
 
 @Component({
     moduleId: module.id,
