+++ src/app/event-view/event-view.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 
 import { Event } from '../shared/event.model';
-import { EventService } from '../shared/event.service';
+import { MockEventService as EventService } from '../shared/mock/mock-event.service';
 import { EventListComponent } from './event-list/event-list.component';
 import { EventDetailComponent } from './event-detail/event-detail.component';
