+++ src/app/shared/mock/mock-event.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import { EVENTS } from './mock-events';
 
 @Injectable()
-export class EventService {
+export class MockEventService {
     getEvents() {
         return Promise.resolve(EVENTS);
     }
